Projekt/requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
python-multipart
psycopg2-binary